New Features - Drag n Drop of playlist & delete tracks from queue.
Updated the main tracklist view to permit deletion of tracks and to
allow drag and drop functionality. Less was updated and rebuilt (despite
Lou working on the new CSS) with extras to enhance the drag and drop
experience.

The main tracklist collection has been abstracted as a permanent object
on the App namespace so that it is no longer recreated/destroyed on page
view changes. Updated the core collection view class to accept an
attribute on new object (defaults to true):
disableCollectionListenersOnRemove. This option will prevent the
collection view from removing event listeners on the collection. Used
here to maintain listeners on the App.tracklist collection.