Issue #16 - Amend .gitignore to include missing build assets!
As part of the update to this project, I incorporated the GitHub
recommended gitignore for Python projects. Unfortunately, it excludes
"lib/" as a default which I missed. This, of course, matched src/lib and
so excluded the application.

In addition, I also spotted a node_module that's not in npm so I've
ensured that it's included here too.